release: automatically update pkg/testutils/release/cockroach_releases.yaml
Previously, as a part of release process we updated this file after the
docs team published the releases yaml file. This is adds a human wait
step in the release process, which is harder to automate.

This PR adds a GitHub Action cronjob to automatically update the file
and create PRs.

Epic: none
Release note: None